requirements.txt
matplotlib>=3.5.0
networkx>=2.6.0
scipy>=1.7.0
pandas>=1.3.0
numpy>=1.20.0
pyyaml>=5.4.0
python-louvain>=0.15
